Rename file for custom config parser
There was a conflict between the name of the module and the name
of the class that was confusing sphinx.
