+++ mpas_analysis/configuration/__init__.py
@@ -1,1 +1,1 @@
-from .MpasAnalysisConfigParser import MpasAnalysisConfigParser
+from .mpas_analysis_config_parser import MpasAnalysisConfigParser
